avoid warnings in rex_file and rex_dir

+++ redaxo/src/core/lib/util/file.php
@@ -52,12 +52,14 @@ class rex_file
    */
   static public function put($file, $content)
   {
-    rex_dir::create(dirname($file));
+    if(!rex_dir::create(dirname($file)) || file_exists($file) && !is_writable($file))
+    {
+      return false;
+    }
 
     if(file_put_contents($file, $content) !== false)
     {
-      chmod($file, rex::getFilePerm());
-      return true;
+      return chmod($file, rex::getFilePerm());
     }
 
     return false;
@@ -104,7 +106,7 @@ class rex_file
         $dstfile = rtrim($dstfile, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR . basename($srcfile);
       }
 
-      if(copy($srcfile, $dstfile))
+      if(is_executable(dirname($dstfile)) && (!file_exists($dstfile) || is_writable($dstfile)) && copy($srcfile, $dstfile))
       {
         touch($dstfile, filemtime($srcfile));
         chmod($dstfile, rex::getFilePerm());
